Merge branch 'supplier/4.14-yocto/SEG_BKC' into release/4.14-yocto/GP_BASE

 Conflicts:
	drivers/media/i2c/crlmodule/crl_adv7481_hdmi_configuration.h
	drivers/media/i2c/crlmodule/crlmodule-data.c
	drivers/media/pci/intel/ipu4/ipu4-css/ipu-wrapper.c
	sound/soc/intel/skylake/skl.c

+++ drivers/dma-buf/hyper_dmabuf/xen/hyper_dmabuf_xen_shm.h
@@ -29,8 +29,8 @@
  * create a table with those in 1st level shared pages then return reference
  * numbers for this top level table.
  */
-int xen_be_share_pages(struct page **pages, int domid, int nents,
-		    void **refs_info);
+long xen_be_share_pages(struct page **pages, int domid, int nents,
+			void **refs_info);
 
 int xen_be_unshare_pages(void **refs_info, int nents);
